requirements.txt
Arpeggio==2.0.3
attrs==25.4.0
debugpy==1.8.17
dill==0.4.0
flake8==7.3.0
grpcio==1.76.0
mccabe==0.7.0
parver==0.5
protobuf==5.29.5
pulumi==3.206.0
pulumi_netbox @ file:///home/matt/Code/pulumi-netbox/sdks/netbox
pycodestyle==2.14.0
pyflakes==3.4.0
PyYAML==6.0.3
semver==3.0.4
setuptools==80.9.0
typing_extensions==4.15.0
wheel==0.45.1
